+++ sdk/storage/azdatalake/file/client.go
@@ -411,13 +411,6 @@ func (f *Client) AppendData(ctx context.Context, offset int64, body io.ReadSeekC
 		return AppendDataResponse{}, err
 	}
 	resp, err := f.generatedFileClientWithDFS().AppendData(ctx, body, appendDataOptions, nil, leaseAccessConditions, cpkInfo)
-	// TODO: check and uncomment this
-	//if err != nil {
-	//	_, err1 := body.Seek(0, io.SeekStart)
-	//	if err1 != nil {
-	//		return AppendDataResponse{}, err1
-	//	}
-	//}
 	return resp, exported.ConvertToDFSError(err)
 }