+++ sdk/storage/azdatalake/file/models.go
@@ -185,6 +185,13 @@ type FlushDataOptions struct {
 	// RetainUncommittedData if "true", uncommitted data is retained after the flush operation
 	// completes, otherwise, the uncommitted data is deleted after the flush operation.
 	RetainUncommittedData *bool
+	// LeaseAction Describes actions that can be performed on a lease.
+	LeaseAction *LeaseAction
+	// LeaseDuration specifies the duration of the lease, in seconds, or negative one
+	// (-1) for a lease that never expires. A non-infinite lease can be between 15 and 60 seconds.
+	LeaseDuration *int64
+	// ProposedLeaseID specifies the proposed lease ID for the file.
+	ProposedLeaseID *string
 }
 
 func (o *FlushDataOptions) format(offset int64) (*generated.PathClientFlushDataOptions, *generated.ModifiedAccessConditions, *generated.LeaseAccessConditions, *generated.PathHTTPHeaders, *generated.CPKInfo, error) {
@@ -230,6 +237,15 @@ func (o *FlushDataOptions) format(offset int64) (*generated.PathClientFlushDataO
 			cpkInfoOpts.EncryptionKeySHA256 = o.CPKInfo.EncryptionKeySHA256
 			cpkInfoOpts.EncryptionAlgorithm = o.CPKInfo.EncryptionAlgorithm
 		}
+		if o.LeaseAction != nil {
+			flushDataOpts.LeaseAction = o.LeaseAction
+		}
+		if o.LeaseDuration != nil {
+			flushDataOpts.LeaseDuration = o.LeaseDuration
+		}
+		if o.ProposedLeaseID != nil {
+			flushDataOpts.ProposedLeaseID = o.ProposedLeaseID
+		}
 	}
 	return flushDataOpts, modifiedAccessConditions, leaseAccessConditions, httpHeaderOpts, cpkInfoOpts, nil
 }
@@ -241,11 +257,22 @@ type AppendDataOptions struct {
 	TransactionalValidation TransferValidationType
 	// LeaseAccessConditions contains optional parameters to access leased entity.
 	LeaseAccessConditions *LeaseAccessConditions
+	// If file should be flushed after append
+	Flush *bool
+	// LeaseAction Describes actions that can be performed on a lease.
+	LeaseAction *LeaseAction
+	// LeaseDuration specifies the duration of the lease, in seconds, or negative one
+	// (-1) for a lease that never expires. A non-infinite lease can be between 15 and 60 seconds.
+	LeaseDuration *int64
+	// ProposedLeaseID specifies the proposed lease ID for the file.
+	ProposedLeaseID *string
 	// CPKInfo contains optional parameters to perform encryption using customer-provided key.
 	CPKInfo *CPKInfo
 }
 
-func (o *AppendDataOptions) format(offset int64, body io.ReadSeekCloser) (*generated.PathClientAppendDataOptions, *generated.LeaseAccessConditions, *generated.CPKInfo, error) {
+func (o *AppendDataOptions) format(offset int64, body io.ReadSeekCloser) (*generated.PathClientAppendDataOptions,
+	*generated.LeaseAccessConditions, *generated.CPKInfo, error) {
+
 	if offset < 0 || body == nil {
 		return nil, nil, nil, errors.New("invalid argument: offset must be >= 0 and body must not be nil")
 	}
@@ -280,11 +307,23 @@ func (o *AppendDataOptions) format(offset int64, body io.ReadSeekCloser) (*gener
 			cpkInfoOpts.EncryptionKeySHA256 = o.CPKInfo.EncryptionKeySHA256
 			cpkInfoOpts.EncryptionAlgorithm = o.CPKInfo.EncryptionAlgorithm
 		}
-	}
-	if o != nil && o.TransactionalValidation != nil {
-		_, err = o.TransactionalValidation.Apply(body, appendDataOptions)
-		if err != nil {
-			return nil, nil, nil, err
+		if o.LeaseAction != nil {
+			appendDataOptions.LeaseAction = o.LeaseAction
+		}
+		if o.LeaseDuration != nil {
+			appendDataOptions.LeaseDuration = o.LeaseDuration
+		}
+		if o.ProposedLeaseID != nil {
+			appendDataOptions.ProposedLeaseID = o.ProposedLeaseID
+		}
+		if o.Flush != nil {
+			appendDataOptions.Flush = o.Flush
+		}
+		if o.TransactionalValidation != nil {
+			_, err = o.TransactionalValidation.Apply(body, appendDataOptions)
+			if err != nil {
+				return nil, nil, nil, err
+			}
 		}
 	}
 
@@ -505,6 +544,8 @@ type SetExpiryValues struct {
 
 // ACLFailedEntry contains the failed ACL entry (response model).
 type ACLFailedEntry = path.ACLFailedEntry
+
+type LeaseAction = path.LeaseAction
 
 // SetAccessControlRecursiveResponse contains part of the response data returned by the []OP_AccessControl operations.
 type SetAccessControlRecursiveResponse = generated.SetAccessControlRecursiveResponse