+++ sdk/storage/azdatalake/internal/path/models.go
@@ -315,3 +315,5 @@ type CPKScopeInfo = blob.CPKScopeInfo
 
 // ACLFailedEntry contains the failed ACL entry (response model).
 type ACLFailedEntry = generated.ACLFailedEntry
+
+type LeaseAction = generated.LeaseAction